refactor: move `extract_file_interface_imports` to cli.utils
